+++ Runtime/Extensions/CollectionExtensions.cs
@@ -2,8 +2,6 @@ using System;
 using System.Linq;
 using System.Collections.Generic;
 
-using Random = UnityEngine.Random;
-
 
 namespace CinderUtils.Extensions {
 
@@ -12,16 +10,27 @@ namespace CinderUtils.Extensions {
 
         public static T GetRandom<T>(this ICollection<T> collection) {
             if (collection.Count == 0) return default(T);
-            else return collection.ElementAt(Random.Range(0, collection.Count));
+            else return collection.ElementAt(UnityEngine.Random.Range(0, collection.Count));
+        }
+
+        public static T GetRandom<T>(this ICollection<T> collection, System.Random RNG) {
+            if (collection.Count == 0) return default(T);
+            else return collection.ElementAt(RNG.Next(0, collection.Count));
         }
 
         public static T LoopingGet<T>(this ICollection<T> collection, int i) {
             if (collection.Count == 0) return default(T);
             else return collection.ElementAt(i % collection.Count);
         }
+
+        public static void AddAll<T>(this ICollection<T> collection, IEnumerable<T> items) {
+            foreach (var item in items) {
+                collection.Add(item);
+            }
+        }
     }
 
-    // Extensions for Collections
+    // Extensions for Dictionaries
     public static class DictionaryExtensions {
 
         public static V GetValue<K, V>(this IDictionary<K, V> dict, K key, V defaultValue = default(V)) {